Use enscons and SCons for build backend
Add custom SCons tools for building wheels and gresource bundles. We are
using enscons just as a PEP 517 shim, even though it comes with wheel
building functionality it's currently sparsely documented. It was not
too much effort to redesign a wheel building tool anyway, and we get
more flexibility this way as well I suppose.