Replace setuptools_scm
setuptools_scm needs to be installed to the same Python as SCons is
installed to (sometimes you might want to install SCons using the
system's package manager instead of pip, and this can make it confusing
to figure out which Python SCons is using). It also does some file path
manipulations could be a problem in cygwin/msys. This adds too many ways
for a build to fail, just from trying to calculate the software version.

This patch replaces setuptools_scm with a simple "git describe" call and
some custom parsing.